backend/services/requirements.txt
Flask==2.3.2
Werkzeug>=2.3.3
flask-jwt-extended==4.4.0  # atau jika perlu, update ke versi terbaru yang kompatibel dengan Flask dan Werkzeug
flask-cors==4.0.0
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0